+++ Sources/RichTextKit/Attributes/RichTextAttribute.swift
@@ -12,3 +12,9 @@ import Foundation
  This typealias represents a rich text dictionary key.
  */
 public typealias RichTextAttribute = NSAttributedString.Key
+
+/**
+ This typealias represents a ``RichTextAttribute`` keyed and
+ `Any` valued dictionary.
+ */
+public typealias RichTextAttributes = [RichTextAttribute: Any]