+++ Sources/RichTextKit/Attributes/RichTextAttributeReader+Font.swift
@@ -11,12 +11,16 @@ import Foundation
 public extension RichTextAttributeReader {
 
     /// Get the font at a certain range.
-    func richTextFont(at range: NSRange) -> FontRepresentable? {
+    func richTextFont(
+        at range: NSRange
+    ) -> FontRepresentable? {
         richTextAttribute(.font, at: range)
     }
 
     /// Get the font size (in points) at a certain range.
-    func richTextFontSize(at range: NSRange) -> CGFloat? {
+    func richTextFontSize(
+        at range: NSRange
+    ) -> CGFloat? {
         richTextFont(at: range)?.pointSize
     }
 }