+++ Sources/RichTextKit/Attributes/RichTextAttributeReader+Style.swift
@@ -11,7 +11,9 @@ import Foundation
 public extension RichTextAttributeReader {
 
     /// Get the text styles at a certain range.
-    func richTextStyles(at range: NSRange) -> [RichTextStyle] {
+    func richTextStyles(
+        at range: NSRange
+    ) -> [RichTextStyle] {
         let attributes = richTextAttributes(at: range)
         let traits = richTextFont(at: range)?.fontDescriptor.symbolicTraits
         var styles = traits?.enabledRichTextStyles ?? []