+++ Sources/RichTextKit/Context/RichTextContext.swift
@@ -266,7 +266,7 @@ public extension RichTextContext {
     /// Set the attributed string to a new rich text.
     func setAttributedString(to string: NSAttributedString) {
         let mutable = NSMutableAttributedString(attributedString: string)
-        mutable.setRichTextFontSize(fontSize)
+        mutable.setRichTextFontSize(fontSize, at: mutable.richTextRange)
         shouldSetAttributedString = mutable
     }